requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
Django==1.11.6
django-map-widgets==0.1.9
idna==2.6
psycopg2==2.7.3.1
pytz==2017.2
requests==2.18.4
urllib3==1.22
